+++ Simulation/Simulation.vcxproj
@@ -144,7 +144,7 @@
     <ClInclude Include="com_Box.h" />
     <ClInclude Include="com_Camera.h" />
     <ClInclude Include="CollisionDetection.h" />
-    <ClInclude Include="Model.h" />
+    <ClInclude Include="com_Model.h" />
     <ClInclude Include="Render.h" />
     <ClInclude Include="Shader.h" />
     <ClInclude Include="Sphere.h" />
@@ -157,7 +157,7 @@
     <ClCompile Include="com_Box.cpp" />
     <ClCompile Include="com_Camera.cpp" />
     <ClCompile Include="CollisionDetection.cpp" />
-    <ClCompile Include="Model.cpp" />
+    <ClCompile Include="com_Model.cpp" />
     <ClCompile Include="Render.cpp" />
     <ClCompile Include="Shader.cpp" />
     <ClCompile Include="main.cpp" />
